Properly terminate Ngrok when NGROK:STOP is called
UIOP:LAUNCH-PROGRAM is now used instead of UIP:RUN-PROGRAM; this enables
us to establish a restart which, when invoked, we can use to gracefully
terminate the running command (interrupting a call to UIOP:RUN-PROGRAM
unfortunately does not cause the underlying command to be terminated).

To actually terminate a running program, one has to invoke
TERMINATE-RUNNING-COMMAND in the same execution context the program we
want to terminate is running on.  This means that the only reasonable
use case for this, is if we had wrapped the RUN call inside
BT:MAKE-THREAD, and now we wanted to break into that running thread
with BT:INTERRUPT-THREAD.  Guess what, that's exactly what NGROK:STOP
is now doing, calling BT:INTERRUPT-THREAD with
TERMINATE-RUNNING-COMMAND.