Unit test NGROK:START and NGROK:STOP (happy path only)
To run the tests, load the tests asd file first:

    > (asdf:load-asd "ngrok-tests.asd")

And then call `(asdf:test-system :ngrok)`.

I tried to have the tests system defined inside the main .asd file, but
the combination of :package-inferred-system, and the use of :pathname,
made it impossible to define a new system with the canonical name
"ngrok/tests" (any system defined inside the tests/ directory, ASDF
would try and load it starting from the main system, "ngrok", and not
the testing one).

That forced me to call the test system something else, i.e.
"ngrok-tests", but then ASDF would complain with the following:

    WARNING: System definition file
    #P"/Users/matteolandi/Workspace/ngrok/ngrok.asd" contains definition
    for system "ngrok-tests". Please only define "ngrok" and secondary
    systems with a name starting with "ngrok/" (e.g. "ngrok/test") in
    that file.

And that's how I ended up creating defining the tests system inside
a .asd file of its own.

Also, it relies on the `NGROK_AUTH_TOKEN` environment variable being
defined (tests actually try to setup a tunnel!) and without it, tests
will fail.